refactor(client): Isolate date controls from dataset card
